First pass at converting shipping class dialog to a fragment
