+++ WooCommerce/src/main/kotlin/com/woocommerce/android/ui/products/ProductShippingClassAdapter.kt
@@ -8,7 +8,7 @@ import android.widget.TextView
 import androidx.recyclerview.widget.RecyclerView
 import com.woocommerce.android.R
 import com.woocommerce.android.ui.products.ProductShippingClassAdapter.ViewHolder
-import com.woocommerce.android.ui.products.ProductShippingClassDialog.ShippingClassDialogListener
+import com.woocommerce.android.ui.products.ProductShippingClassFragment.ShippingClassFragmentListener
 import kotlinx.android.synthetic.main.product_shipping_class_item.view.*
 import org.wordpress.android.fluxc.model.WCProductShippingClassModel
 
@@ -16,7 +16,7 @@ import org.wordpress.android.fluxc.model.WCProductShippingClassModel
  * RecyclerView adapter which shows a list of product shipping classes, the first of which will
  * be "No shipping class" so the user can choose to clear this value.
  */
-class ProductShippingClassAdapter(context: Context, private val listener: ShippingClassDialogListener) :
+class ProductShippingClassAdapter(context: Context, private val listener: ShippingClassFragmentListener) :
         RecyclerView.Adapter<ViewHolder>() {
     companion object {
         private const val VT_NO_SHIPPING_CLASS = 0
