+++ WooCommerce/src/main/res/navigation/nav_graph_main.xml
@@ -254,5 +254,12 @@
             android:name="remoteProductId"
             android:defaultValue="0L"
             app:argType="long" />
+        <action
+            android:id="@+id/action_productShippingFragment_to_productShippingClassFragment"
+            app:destination="@id/productShippingClassFragment" />
     </fragment>
+    <fragment
+        android:id="@+id/productShippingClassFragment"
+        android:name="com.woocommerce.android.ui.products.ProductShippingClassFragment"
+        android:label="ProductShippingClassFragment" />
 </navigation>
